app/requirements.txt
atproto>=0.0.48
atproto-client>=0.0.44
httpx>=0.27.0
PyYAML>=6.0.1
tenacity>=8.2.3
uvloop>=0.17
